added chapters for day 1&2 and day1 exercise

+++ basics/functions.py
@@ -11,4 +11,4 @@ def stretch(image):
 
         temp[:,:,i] = (image[:,:,i]-minI)*(((maxO-minO)/(maxI-minI))+minO)
     
-    return np.float32(temp)
+    return temp
